refactor: package product dto
