+++ src/main/java/com/example/repick/domain/product/service/ProductService.java
@@ -1,6 +1,7 @@
 package com.example.repick.domain.product.service;
 
-import com.example.repick.domain.product.dto.*;
+import com.example.repick.domain.product.dto.product.*;
+import com.example.repick.domain.product.dto.productOrder.GetProductCart;
 import com.example.repick.domain.product.entity.*;
 import com.example.repick.domain.product.repository.*;
 import com.example.repick.domain.product.validator.ProductValidator;
