+++ src/main/java/com/example/repick/domain/user/api/UserController.java
@@ -218,4 +218,6 @@ public class UserController {
         return SuccessResponse.success(userService.verifySmsVerification(postVerifySmsVerification));
     }
 
+
+
 }
